requirements.txt
# Core dependencies
aiohttp==3.9.1
asyncio==3.4.3
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.1
elasticsearch==8.11.0
praw==7.7.1

# NLP and ML
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
langdetect==1.0.9
rake-nltk==1.0.7

# Utilities
pydantic==2.5.2
python-dateutil==2.8.2
pytz==2023.3

# API framework
fastapi==0.104.1
uvicorn==0.24.0

# Monitoring
prometheus-client==0.19.0

